Refactoring so preload script don't call window
