+++ app/components/BakerPendingChange/BakerPendingChange.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { BakerPendingChange } from '@concordium/node-sdk';
+import type { BakerPendingChange } from '@concordium/node-sdk';
 import { useConsensusStatus } from '~/utils/dataHooks';
 import { displayAsGTU } from '~/utils/gtu';
 import { epochDate, getFormattedDateString } from '~/utils/timeHelpers';
